Rearrange .bash_profile; add mvim aliases.
Move colors around in .bash_profile. Move heroku restart to
.bashrc_MacBookPro

Add hr, vi, and vim aliases that use the mvim command. This is better
than creating symbolic links in the /usr/local/bin directory because
it's in version control and does not require symbolic links to absolute
locations on the computer (if mvim works, vi and vim works).